Redid primary and sensitivity analyses #1 using appropriate residuals
In this doc I calculated three different sets of residuals for each BA measure. The first represents the previous versions, calculated in the sample with full biomarker data (N=5,870). The second is calculated within the subset of participants that make up the primary sample (N=4,418). The third is calculated within the subset of participants that make up the sensitivity analysis sample (N=5,184). I did this because it felt inappropriate to use residuals that were based on minimizing error in a different sample than the one being used for the given analysis. That said, literally only two digits changes across all primary and sensitivity analysis 1 models. After seeing this I did not rerun construct new residuals for sensitivity analysis 2 or 3.